KNL-2
Migrated the test harness to the kernel and migrated the tests.
There are some test failures that need investigating, the build is configured to run anyway as at the moment I believe
that the test failures are as a result of the tests and not as a result of the code which they are testing.

This is not a complete solution since it does not at present load each component into its own classloader and so will no
validate for cross classloader issues, but it does ensure that the classpath is fully defined in in the poms and so all
the tests run with mvn clean install and nothing else. The tests also run in eclipse correctly and its releatively easy
to bring up a new test case.

I am not certain if Sakai core wants to use the mecahnism here, but if they do the TestComponentManagerContainer is the class that
enables a test to be created.

Unfortunately because of the excessive number of statics in the framework tests must all be run in a pertest fork giving each
test a new JVM. Even then there are some things that will break the tests including the setting of sakai.home to a specific location.
If you write a test that does this, you MUST set it back to null on teardown.


git-svn-id: https://source.sakaiproject.org/svn/kernel/trunk@51118 66ffb92e-73f9-0310-93c1-f5514f145a0a