KNL-31
Patch by Ray Davis
Sakai code often needs to refresh DB-cached site and group memberships to match the current state of a group provider. (For a recent example, see SAK-14336.) Currently, the Kernel API provides no direct way to do this, and doesn't document an indirect way.

It might be ideal to expose the current BaseAuthzGroupService.Storage refreshAuthzGroup method in the public AuthzGroupService interface. As a minimal stopgap, however, the attached patch file just adds a note to the AuthzGroupService JavaDoc.

Thanks


git-svn-id: https://source.sakaiproject.org/svn/kernel/trunk@52429 66ffb92e-73f9-0310-93c1-f5514f145a0a