SAK-14640 - Added comma-separated support to property parsing.
Note that values may not be quoted or contain commas with this syntax. To use
commas within values, the legacy syntax (.count) must be used. It is possible
to add full CSV support, but Commons Configuration does not include it. It
may be that using Commons CSV is a better overall approach, since we are
already managing the file reading and settings storage.



git-svn-id: https://source.sakaiproject.org/svn/kernel/trunk@106440 66ffb92e-73f9-0310-93c1-f5514f145a0a