KNL-1011 - Fix extra user-site cache invalidation
There was an extra call to invalidate user-site caches for all members
of a site upon calling cacheSite, which lead to significant stomping of
the values. There are event observers to catch site updates, which
handle invalidation when membership changes. This is the appropriate
time as opposed to clearing upon caching of each site.



git-svn-id: https://source.sakaiproject.org/svn/kernel/trunk@123981 66ffb92e-73f9-0310-93c1-f5514f145a0a