KNL-1294 Fix getSites() when used passed joinable and properties.
If you attempted to search for joinable sites and also supplied properties to search for the parameters would be subsituted in the wrong place in the SQL resulting in no results returned (or the wrong results). This fixes the order that the values array is built up in so it matches the SQL string.

Original patch from Angel Nudea Lozano.

git-svn-id: https://source.sakaiproject.org/svn/kernel/trunk@315629 66ffb92e-73f9-0310-93c1-f5514f145a0a